backend/requirements.txt
# Core AI & LangChain
langchain==1.2.10
langchain-core==1.2.17
langchain-google-genai==4.2.1
langchain_google_vertexai
google-cloud-aiplatform==1.139.0

# Web Framework
fastapi==0.135.1
uvicorn[standard]==0.34.0
#pydantic>=2.10.0
pydantic-settings==2.13.1
python-multipart

# Google Cloud Services
google-cloud-vision==3.7.2
google-cloud-bigquery==3.29.0
google-cloud-documentai>=3.0.0
google-cloud-aiplatform>=1.38.0
pillow>=10.0.0

